Revert all the changes made to ensure that children don't get processed > 1 time.
This issue (demonstrated in https://github.com/balderdashy/waterline/issues/1520) is related to the fact that in some cases (singular but NOT plural associations), populated records are added to the parent by reference.  A better solution will be to fix that behavior (using a shallow clone), since it could cause other issues down the road anyway (for example, in userland it would be unexpected that changing a deep value in a returned record could affect another returned record).